# lib/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "JustMeat API",
        "description": "A minimal and easy create a CRUD style API in NodeJs using TypeScript"
    },
    "servers": [
        {
            "url": "/",
            "description": "localhost:3006"
        }
    ],
    "host": "localhost:3006",
    "basePath": "",
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "JWT": {
            "description": "",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "description": "This is the authentication token you need to insert:  ",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "JWT": []
        }
    ],
    "paths": {
        "/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get all restaurants in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Restaurants"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Create Restaurant",
                "description": "Create a new restaurant",
                "requestBody": {
                    "description": "Create restaurant",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Restaurant"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant Create"
                    },
                    "422": {
                        "description": "Empty Array"
                    },
                    "403": {
                        "description": "Name already in use"
                    }
                }
            }
        },
        "/restaurants/{id}": {
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Edit Restaurant by ID",
                "description": "This can only be done by the logged in user",
                "operationId": "modifyRestaurantById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of subject to return",
                        "required": true,
                        "style": "simple",
                        "explode": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Modify new subject",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Restaurant"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": ["Restaurants"],
                "summary": "Delete Restaurant",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteRestaurant",
                "produces": ["application/json"],
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "The restaurant that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }],
                "responses": {
                    "400": {
                        "description": "Invalid restaurant supplied"
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Restaurant": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "plates": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "price": {
                                "type": "number"
                            }
                        }
                    }
                },
                "rating": {
                    "type": "number"
                },
                "typology": {
                    "type": "string"
                }
            }
        },
        "Restaurants": {
            "type": "object",
            "properties": {
                "restauants": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#definitions/Restaurant"
                    }
                }
            }
        },
...
    }
}